💄: Avoid flashing background image when navigating screens
