+++ example-app/SantokuApp/android/app/src/main/java/jp/fintan/mobile/santokuapp/AppActivity.java
@@ -1,0 +1,44 @@
+package jp.fintan.mobile.santokuapp;
+
+import android.os.Bundle;
+
+import com.facebook.react.ReactActivity;
+import com.facebook.react.ReactActivityDelegate;
+import com.facebook.react.ReactRootView;
+import com.swmansion.gesturehandler.react.RNGestureHandlerEnabledRootView;
+
+import expo.modules.splashscreen.singletons.SplashScreen;
+import expo.modules.splashscreen.SplashScreenImageResizeMode;
+
+public class AppActivity extends ReactActivity {
+  @Override
+  protected void onCreate(Bundle savedInstanceState) {
+    // https://github.com/software-mansion/react-native-screens#android
+    // https://github.com/software-mansion/react-native-screens/issues/17
+    // Discard any Activity state persisted during the Activity restart process, to avoid inconsistencies that lead to crashes.
+    super.onCreate(null);
+    // SplashScreen.show(...) has to be called after super.onCreate(...)
+    // Below line is handled by '@expo/configure-splash-screen' command and it's discouraged to modify it manually
+    SplashScreen.show(this, SplashScreenImageResizeMode.NATIVE, ReactRootView.class, true);
+  }
+
+
+    /**
+     * Returns the name of the main component registered from JavaScript.
+     * This is used to schedule rendering of the component.
+     */
+    @Override
+    protected String getMainComponentName() {
+        return "main";
+    }
+
+    @Override
+    protected ReactActivityDelegate createReactActivityDelegate() {
+        return new ReactActivityDelegate(this, getMainComponentName()) {
+            @Override
+            protected ReactRootView createRootView() {
+                return new RNGestureHandlerEnabledRootView(AppActivity.this);
+            }
+        };
+    }
+}
